+++ packages/core/src/js/integrations/default.ts
@@ -89,7 +89,7 @@ export function getDefaultIntegrations(options: ReactNativeClientOptions): Integ
     if (options.attachViewHierarchy) {
       integrations.push(viewHierarchyIntegration());
     }
-    if (options._experiments && typeof options._experiments.profilesSampleRate === 'number') {
+    if (typeof options.profilesSampleRate === 'number') {
       integrations.push(hermesProfilingIntegration());
     }
   }